[NW] Creates calendaring helper
